pyproject.toml
[project]
name = "btc-agent-lightning"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "langchain-anthropic==0.3.10",
    "langchain-community==0.3.20",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-openai==0.3.9",
    "langgraph==0.3.18",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langgraph-cli[inmem]==0.1.79",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/ehallmark/btc-agent-lightning"
Issues = "https://github.com/ehallmark/btc-agent-lightning/issues"
